BLXV-69 (8.36) Add support for multiple filewriters in additionalPluginList.
The SingleImagePerFileWriter, MultipleImagesPerParallelHDF5FileWriter and
MultipleImagesPerHDF5FileWriter classes rely on the default implementation
of getName() in FileWriterBase. This means that you cannot add more than
one filewriter to the additionalPluginList of an NXDetector.

By moving the implementation of getName() out of abstract FileWriterBase
class and into the direved classes, different values can be returned for
each type of filewriter, allowing them to be used simultaneously.