EPICS (8.36) Revert commit 78c65d0 to prevent failure when LazyOpen unsupported
Although this fixed the problem where Area Detector IOC was restarted
before the GDA server, it did not fix the case where the IOC was restarted
*after* the.

The real fix is to add a <property name="lazyOpen" value="true" /> to
the MultipleImagesPerHDF5FileWriter bean for each detector which supports
LazyOpen. This will cause LazyOpen to be set in every call to
prepareForCollection.