BLXV-137 (8.38) Restore setImageNDArrayPortInput into ADControllerBase.
This appears to have been unintentionally removed in commit cd0ca22 during
a refactor.

Without it the GDA client often resets the array plugin input to be the
camera driver. This is a problem for detectors such as the Perkin Elmer
on I15, which need processing to be performed on images before they are
displayed.