Revert "avoid 'g.d.m.EpicsMotor - Error: illegal .{HLM,LLM} value.' message if dbr.isSHORT; use motor name in message otherwise"
This reverts commit 9c644a34e9fd0f255e70d988d520f05282d8e7b8 which
caused .HLM and .LLM values on I04-1 to be set to 0. It transpires that
they are initially set to the correct values via
EpicsMotor.set{Max,Min}PositionFromListener(double {max,min}Position).
The problem is that EpicsMotor.{HLM,LLM}MonitorListener implementations
log the titular message in the first place, as dbr.isSHORT() returns
true, contradicting cainfo on the PVs!