BLXV-195 (8.44) Fix race condition with Cached PVs and RBVs.
Where a PV pair of PVs are used to set and read back a value, and you
read back soon after setting, you may end up with the old cached value
instead on the new value. To prevent this, cached readback values need to
be reset whenever the PV setting them is set.

Change-Id: I2a22c2ec5414d1a6065c54a30f2fd3f4faabc4bf