BLXV-195 (8.44) Create pluggable collection strategies and decorators.
This collection strategy system is intended to replace the monolithic
collection strategies in gda.device.detector.addetector.triggering with
a set of reusable collection strategies and decorators which maximise
flexibility while removing the need for the same code to be copied into
multiple all classes which need that functionality.

This allows us to apply the same set of collection strategy decorators
(with the same behaviour) to multiple collection strategies.

Change-Id: Iffccd8a34b3f845e6d3b79155a639f951714bb18