I20-79 : New Eurotherm furnace scannable
This class controls a Eurotherm903 furnace using the EPICS interface. 
It was originally written for use with the B18 Eurotherm (in a B18 specific plugin). This is a cleaned up version of that class developed when using the same furnace on I20 with 'moveAndWait' function added, and moved to a non beamline specific package.

Change-Id: I96d69144c6bc691dda4085cdb2115a3645652c03