BLXVIII-202 Prevent exception when position is null
When calling getPosition() outside the context of a scan, the "pos"
Object[] is not null itself, but is a one-element array containing a
null reference. This is returned and causes ScannableUtils to throw an
exception when it tries to format the position.

After this change, the null position is identified and we fall back to
super.getPosition(), which calls getPosition() on the underlying motor
and therefore is much more likely to return a sensible value.

Change-Id: I174389468bbf5ab47f489656d3b947b0273ee8b7