DAQ-224 Make DummyRealPositionReader extend ScannableMotionUnitsBase
This makes it as similar as possible to the EpicsTrajectoryScanRealPositionReader
class and prevents exceptions in I18's localStation.py caused by the
real version implementing Scannable (and therefore being added automatically
to the Jython namespace) and the dummy version not.

We have to override rawGetPosition() to avoid exceptions when 'pos' is
called.

Change-Id: Iee7d0840d972ffbdd65f3d5296fb9810a16611a8