B07-76 Implement a class for controlling a Femto+Bekhoff ADC pair
These are in use on several beamlines and the existing solutions are
limited and can't be used on b07 as the Femto's there support two
operation modes. Additionally previous implementations don't trigger the
ADC and therefore the data is not guaranteed to be correct as was
observed on b07. 

This implementation supports setting the averaging time in a way
consistent with other detectors, and writes a richer data into the
NeXus file for further analysis.

Change-Id: Iabd77f52ab218866938d2ba544d2a6ede8094a4d