Sort regions list in PlotServerROISelectionProvider before returning
We want the region order to be preserved. Regions are usually returned
in creation order, *except* the currently selected region, which is
always returned first. This is inconvenient when we want a consistent
ordering for indexing purposes.

Sorting by the name of a region is imperfect, but at least *mostly*
consistent (ignoring region renames), but that shouldn't matter.

Change-Id: Ia033b4725e8724beb4810b8f95d2df5176747ef9