[DAQ-1128] New interface containing functions needed by EDXDController
The interface XmapEpicsDevice extends IEpicsDevice and contains
the functions that EDXDController and its associated classes
need to control an XMAP device.

FindableEpicsDevice now inherits from this, so this can continue to
be used in live mode and a dummy object implementing the same
interface can be created.

For convenience, added a new function to FindableEpicsDevice to get
a single PV, rather than returning the PV map

Change-Id: I58ef61603aa1237607df451ad4287699a7ae79e8