I20_1-197 : Changes to TrajectoryScan classes to add methods to interface
Changes to trajectory scan controller classes so they can be used in I20-1's
TrajectoryScanPreparer class to provide PV access. These changes are based
on current tested functionality in TrajectoryScanPreparer.
- New methods added to interface (mainly 'getters' to complement the 'setters').
- 'sendProfiles' method moved to base class
- New methods added for building up trajectory by appending points.
- Dummy controller updated to implement new methods on interface;
arrays passed to dummy controller that are sent to Epics in live controller
are now stored in arrays to facilitate unit tests.

Change-Id: Ifab570247f09cf74a9eabe91bb93618d98a39e55