DAQ-1481 Fix race condition in EpicsChannelManager.createChannel
An EpicsChannelManager is often created with an InitializationListener.
EpicsChannelManager will call the listener's initializationCompleted method
when:

 1. The class using the EpicsChannelManager has indicated that it has finished
    requesting channels - done by calling
    EpicsChannelManager.creationPhaseCompleted(), which sets the
    creationPhaseCompleted flag; AND

 2. All "critical" (non-optional) channels are connected.

To keep track of whether a critical channel is connected:

 1. createChannel calls EpicsController.createChannel to create a Channel
    object, and then adds the Channel object to unconnectedCriticalChannels.

 2. When the channel becomes connected, EpicsChannelManager.connectionChanged
    is called, which removes the channel from unconnectedCriticalChannels.

There's a race condition here, because it's possible for the channel to become
connected (and for EpicsChannelManager.connectionChanged to try to remove it
from unconnectedCriticalChannels) before EpicsChannelManager.createChannel has
added it to unconnectedCriticalChannels.

That will result in unconnectedCriticalChannels having channels added to it
that subsequently never get removed. As a result, initializationCompleted is
never called.

Fix this by adding the channel to unconnectedCriticalChannels *before* calling
EpicsController.createChannel.

Since we don't have a reference to a Channel object before
EpicsController.createChannel is called, change unconnectedCriticalChannels
from a Set of Channel objects to a Set of Strings (channel names).

Change-Id: I1d6699a51ba963f9194e86a89304388614291c93