I15-351 (9.8) Terminate endRecording() loop when captured frames exceed expected
When streaming with setNumCapture set to -1 (forever) the HDF5 plugin will
not stop itself, so we have to also check whether we have captured as many
frames as we expected.

This fix enables the loop to terminate even when the number of expected
frames is not known and thus defaults to zero. This can happen when using
a MultiRegionScan where each individual region scan knows only about it's
own frames.

Since it appears that endRecording doesn't get called until all of the
read() calls have completed, we already know that the HDF writing is
complete, thus we can safely stop the HDF writer capturing.

Originally live on beamline on branch gda-9.8 and cherry picked to master

Change-Id: Id5bd283dfa45d0c33d44afe67195f9ce4b5373ec