VMXI-432 - Add pneumatic control class that checks limit switches
There are some IN/OUT pneumatics on VMXI that do not follow the typical
pneumatic or valve control that operate via epics callbacks, but
strike limit switches when they reach the target position.

A separate issue also prevents the use of CA callbacks being used on
one of the devices (VMXI-219).

Change-Id: Ic231d6a7b75774e79d84273f01e5c8cf24e52d28