[DAQ-2173] EpicsMonitor: fix testing of notifications
If the sensitivity of the monitor is set, not every change in the
value of the monitor is notified. In deciding whether to notify
observers of a particular event, what is important is therefore
whether the latest value differs sufficiently from the last value
that was *reported*, not the last value received.

This change records the last reported value and uses this in deciding
whether to notify observers.

This also fixes what looks like bugs in handling arrays:

(1) For numerical arrays, bring the test into line with the simple
value case i.e. (latest - last_reported) / last_reported
rather than (latest - last - 1)

(2) For String and byte arrays, clone the previous-value-received
array before reassigning it.

Change-Id: I5dcda5f06df3be2038e0689dd39e1a38bf1b6c46