DAQ-2231 Ensure EDXDControllerQMap configures itself only once
Creating a scannable and adding it to a ScannableGroup (as is common on beamlines)
results in the scannable's configure() function being called twice, as
ScannableGroup calls configure() on all its members.

This was causing a particular problem for the I12-specific EDXDControllerQMap, as
the base class configure() adds all the subdetectors to a list, and this was causing duplication.

This fixes the issue for GDA 9.13 I12. The equivalent fix for the base class EDXDController
(DAQ-2230) is better put in master, as it is not causing an immediate problem on the beamlines
which use EDXDController directly.

Change-Id: I2b982ab55e7e344c74dfa128b307ad77b418870c