B18-406 : BufferedNXDetector : use 'current point' PV from ADRoiStatsPair plugin
This change enables BufferedNXDetector to use the same PV for the 'current point'
count as used by the NXDetector by extracting it from the ADRoiStatsPair plugin.
The old manually set value 'TSCurrentPoint' was not correct when using new IOC
which requires 'TS:TSCurrentPoint'. The PV name to use can also be manually
set using 'setCurrentPointPointPvName' if required.

Change-Id: Ibbbe1ccc0099e50a5c6ae5b491cfee33767bd7b1