DummyEpicsMonitor: enable setting of array values in Spring
Spring (at least the version we have at the moment) does not allow
you explicitly to set an array value. You can specify a <list>, which
will be converted into an array if necessary, but DummyEpicsMonitor
takes an Object, so Spring will make this conversion.

This change modifies DummyEpicsMonitor.setValue() to convert a List
to an array.

Change-Id: Ib23c87af787ab027f0c8faca07d43775b8092a7a