MXGDA-3344 Remove LazyPV type mismatch warning
There are a number of instances where either GDA have mapped PV's
incorrectly to the wrong data types or where EPICS have configured the
PV's to be of the incorrect types.  As there is little possibility of
Controls checking and updating the types of all of their PV's, the only
other option would be for GDA to check all of the PV's causing this
warning to be produced and working around the issue.
As GDA copes with the mis-mapping the most pragmatic solution to
removing the warning is to stop them from being produced.  If Controls
do ever review their PV mappings then this warning can be reinstated
so that GDA can fix any remaining mis-mappings.

Change-Id: I7971d7282207d7cad9d45875a954ed4e05f3a87c